refactor(ff-decode): extract resampling logic into audio/resample_inner.rs
audio/decoder_inner.rs (1542 lines) mixed codec setup, packet decoding,
seeking, format conversion, and SwResample conversion in one file.

Extract the resampling / frame-conversion concern into a new module:

- audio/resample_inner.rs (~553 lines)
  - SwrContextGuard (RAII for SwrContext)
  - convert_sample_format / sample_format_to_av (pub(crate) format helpers)
  - create_channel_layout (private; used only by the swr pipeline)
  - extract_planes / av_frame_to_audio_frame (AVFrame → AudioFrame)
  - convert_frame_to_audio_frame (dispatcher: swr or direct path)
  - convert_with_swr (full libswresample pipeline)
  - 4 unit tests covering format-mapping round-trips

- audio/decoder_inner.rs reduced to ~1152 lines
  - Removed dead swr_ctx field (was always None; SwrContext is now
    created and freed per-frame inside resample_inner)
  - Call sites updated to use resample_inner::convert_frame_to_audio_frame
    and resample_inner::convert_sample_format

Public API (AudioDecoder) is unchanged.

Closes #717
